Add ft_memcmp

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: gozon <gozon@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/21 11:36:12 by gozon             #+#    #+#             */
-/*   Updated: 2024/05/21 13:22:52 by gozon            ###   ########.fr       */
+/*   Updated: 2024/05/21 14:06:46 by gozon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] && s2[i] && i < n - 1 && s1[i] == s2[i])
